Add check for required attrs in CreateMixin
Add check for required attributes in the `transip.mixins.CreateMixin`
when creating an API object.

This commit also stores the domains tests data in
`tests/fixtures/*.json` to make the test code more readable. It also
uses a `unittest.TestCase` class to prevent duplicate responses and
prepare compatibility with the default `unittest` library.

Signed-off-by: Roald Nefs <info@roaldnefs.com>