Add option to authentication using private key
The TransIP API also allows the creation of an access token by using a private key. By allowing the user to specify either an access token or private key when initialising a new transip.TransIP client, the client could dynamically generate a new access token when a private key is used.

Fixes #14

Signed-off-by: Roald Nefs <info@roaldnefs.com>